LLVM instructions: unary operations, templates for binary operations
