+++ src/llvm/type_system/instructions/binary_operations.rs
@@ -1,0 +1,9 @@
+//! # Binary Operations
+//!
+//! Binary operators are used to do most of the computation in a program.
+//! They require two operands of the same type, execute an operation on
+//! them, and produce a single value. The operands might represent multiple
+//! data, as is the case with the vector data type. The result value has
+//! the same type as its operands.
+//!
+//! https://llvm.org/docs/LangRef.html#binary-operations
